+++ src/zope/configuration/config.py
@@ -333,13 +333,15 @@ class ConfigurationMachine(ConfigurationAdapterRegistry, ConfigurationContext):
     includepath = ()
     info = ''
 
-    #: These exceptions are allowed to be raised from `execute_actions`
-    #: without being re-wrapped into a `ConfigurationExecutionError`.
+    #: These `Exception` subclasses are allowed to be raised from `execute_actions`
+    #: without being re-wrapped into a `ConfigurationExecutionError`. (`BaseException`
+    #: instances are never wrapped.)
+    #:
     #: Users of instances of this class may modify this before calling `execute_actions`
     #: if they need to propagate specific exceptions.
     #:
     #: .. versionadded:: 4.2.0
-    pass_through_exceptions = (KeyboardInterrupt, SystemExit)
+    pass_through_exceptions = ()
 
     def __init__(self):
         super(ConfigurationMachine, self).__init__()
@@ -390,7 +392,7 @@ class ConfigurationMachine(ConfigurationAdapterRegistry, ConfigurationContext):
                     callable(*args, **kw)
                 except pass_through_exceptions:
                     raise
-                except:
+                except Exception:
                     t, v, tb = sys.exc_info()
                     try:
                         reraise(ConfigurationExecutionError(t, v, info),