+++ CHANGES.rst
@@ -52,6 +52,11 @@ Changes
   ``ConfigurationExecutionError``. See `issue 10
   <https://github.com/zopefoundation/zope.configuration/issues/10>`_.
 
+- Stop catching ``BaseException`` and wrapping it in either
+  ``ConfigurationExecutionError`` or ``ZopeXMLConfigurationError``.
+  ``SystemExit`` and ``KeyboardInterrupt`` were always allowed to
+  propagate; now ``GeneratorExit`` and custom subclasses of
+  ``BaseException`` are also allowed te propagate.
 
 4.1.0 (2017-04-26)
 ------------------